Renamed sp and tm regions
    renamed:    src/nupic/regions/SPRegion.py ->
src/nupic/regions/sp_region.py
    renamed:    src/nupic/regions/TMRegion.py ->
src/nupic/regions/tm_region.py
