+++ examples/network/network_api_demo.py
@@ -29,8 +29,8 @@ from pkg_resources import resource_filename
 from nupic.data.file_record_stream import FileRecordStream
 from nupic.engine import Network
 from nupic.encoders import MultiEncoder, ScalarEncoder, DateEncoder
-from nupic.regions.SPRegion import SPRegion
-from nupic.regions.TMRegion import TMRegion
+from nupic.regions.sp_region import SPRegion
+from nupic.regions.tm_region import TMRegion
 
 _VERBOSITY = 0  # how chatty the demo should be
 _SEED = 1956  # the random seed used throughout
