+++ src/nupic/regions/sp_region.py
